Blacklist kombu-4.6.4
Commit [1] bumped the upper constraint for kombu to 4.6.4. Per issues
[2][3] this interacts poorly with eventlet and breaks many things
openstack on py37 with this error:

TypeError: open: path should be string, bytes or os.PathLike, not _NormalAccessor

It is possible there will be a simple way to fix those many things, but
in the meantime, this commit blacklists 4.6.4 to get the world moving
again.

Thanks to Daniel Leaberry for doing the legwork to identify the problem
[4].

[1] Id71496bcb5c8342851f287e35931431e8a443d23
[2] https://github.com/eventlet/eventlet/issues/534
[3] https://github.com/nameko/nameko/issues/655
[4] http://lists.openstack.org/pipermail/openstack-discuss/2019-August/008553.html

Change-Id: I113c06f6e09727aa279f5ab8d2b7534a019394c2