Update patch set 1
Patch Set 1:

> so this is a temporary thing?  if so, how temporary?  I ask because
 > the freeze is coming up in a couple of days

That's a good question. I expect fixing the issues to get 1.7.0 unblocked to be relatively trivial, but I haven't dug into it in depth. I'm also not sure if I'll get time to do that refactor within the next couple of days.

So I guess it doesn't sound all that temporary for Train.

Patch-set: 1