Add two StorPool libraries for the Cinder StorPool driver.
* Is the library actively maintained?

Yes, both libraries are maintained by StorPool. The "storpool" library
is actually part of the StorPool distributed storage software, it is
installed in each deployment since the StorPool tools use it.
The "storpool.spopenstack" library is installed as part of each
deployment of StorPool on an OpenStack controller or hypervisor.

* Is the library good code?

We would certainly like to think so. Both libraries are open-source and
we would welcome reviews and comments.

* Is the library python 3 compatible?

Yes, the StorPool third-party CI at https://spfactory.storpool.com/ runs
the Cinder-related Tempest tests using Python 3.7 on Ubuntu 16.04
(Xenial).

* Is the library license compatible?

Yes, both libraries are under an Apache 2.0 license.

* Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

Not yet, but we have WIP packages that we already install at some of our
customers' locations. We intend to add them to Debian and, hence,
Ubuntu, and we will offer our full assistance to anyone who wants to
officially package them for Fedora.

* Is the function of this library already covered by other libraries in
global-requirements.txt?

No; these libraries are closely related to Python language bindings for
the StorPool distributed storage system API, there is no other package
that offers similar functionality.

* Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

The StorPool volume driver within Cinder uses both the storpool and
storpool.spopenstack libraries. The goal of adding them to
global-requirements is to subsequently add them to Cinder's setup.cfg
file in the [extras] section, so people who want to install the Cinder
Python library and have the StorPool driver work will have them pulled
in as needed. See also https://review.opendev.org/680530

* If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?

N/A

Change-Id: I947c5a0087225668523e4aa249b1b11ef9b850e9