Adding niet.
* Is the library actively maintained?

Yes, this module is actively maintained.

* Is the library good code? We would certainly like to think so.

This module is open-source and we would welcome reviews and comments.

* Is the library python 3 compatible?

Yes, the third-party CI at https://travis-ci.org/openuado/niet runs the
tests using Python 3.7 on Ubuntu 16.04 (Xenial).

* Is the library license compatible?

Yes, this module is under a MIT license.

* Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Not yet. We intend to add them to Debian and, hence, Ubuntu, and also for Fedora.

* Is the function of this library already covered by other libraries in
global-requirements.txt?

No; there is no other package that offers similar functionality.

* Is the library required for OpenStack project or related dev or
infrastructure setup?

Not yet; The goal of adding this to global-requirements is to subsequently add
niet usage to shell scripts to be able to use its powerful features.

* If the library release is managed by the Openstack release process does it
use the cycle-with-intermediary release type?

N/A

Change-Id: I2117fd493ea7216f1a804a8abc544a4dddbba3ac