Add new dependency for akamai API library
* Is the library actively maintained?

This library has not a lot of commits per day, because implements
Authentication handler for requests that provides the Akamai {OPEN}
Edgegrid Authentication scheme to work with Akamai API.
So it's stable enough and last merged pull request was in this year
(28 Feb 2019).

* Is the library good code?

Yes. It's simple enough and has a batch of tests for the existing code.
https://github.com/akamai/AkamaiOPEN-edgegrid-python/tree/master/akamai/edgegrid/test

* Is the library python 3 compatible?

Yes. According the history it has necessary support:
it's support pytho3 since 1.0.4 version
https://github.com/akamai/AkamaiOPEN-edgegrid-python/blob/master/HISTORY.rst#104-2014-11-05

* Is the library license compatible?

Yes, library is under an Apache 2.0 license.

* Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

It has not package for Ubuntu, but there is a package for Fedora.
https://koji.fedoraproject.org/koji/packageinfo?packageID=28118

* Is the function of this library already covered by other libraries in
global-requirements.txt?

There is no alternative libraries can be used to authenticate with Akamai Edge.

* Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

It's used for implementation one of backends for Designate project.
Moreover akamai fast dns api v1 is deprecated, therefore a new back-end
driver is needed in Openstack Designate to support akamai fast dns v2.

* If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?

N/A

Change-Id: Id0a8de7917bb91608b16953ed63722ac2d061ac3