Update patch set 1
Patch Set 1: -Code-Review

Okay, good that this passed since the heat change merged, but I want to think through what that means from the perspective of the requirements repo.

That is, shouldn't we need to be proposing a bump to upper-constraints that pulls in the new release of heat in order for this to work?

I suspect (but haven't confirmed) that this passes because the job is set up to pull in master tips. So we may want to override it to have an empty required_projects...

Patch-set: 1