Update patch set 1
Patch Set 1: -Code-Review

My concern is this:

We run CI against $random_change. It fails in heat because $random_change exposed a heat bug. We fix heat in master. CI passes. We let $random_change go through. But $random_project now starts failing *its* CI because the *released* heat -- against which $random_project is running, at least for some jobs, because heat isn't considered a "required project" -- is still buggy.

From the perspective of openstack/requirements, this seems like exactly the reason to make sure the job is pinned to real released-and-upper-constrained (and even lower-constrained??) versions of the projects it touches.

I could be way off base here, so I'll at least remove my -1, but (obviously) we need requirements cores to weigh in anyway.

Patch-set: 1