Update patch set 1
Patch Set 1:

The problem there is that we don't have releases tracked in reqs (on purpose) for top level projects.

Projects don't do releases often enough.
It encourages using projects as libraries (This is me staring at neutron...)

What is generally recommended is that projects actually follow deprication warnings and remove depending on features that will be pulled.  If projects don't do that then this will happen even if we do pull in tags.

Patch-set: 1