Update patch set 1
Patch Set 1: Code-Review+1

I'm really on the fence with this.  It totally fixes a bug but that code is just wrong and we have a much better way to handle things now.

We can just make the script actually require openstack_requirements and treat it like any other library rather than doing this magic dance.

Patch-set: 1