Update patch set 1
Patch Set 1: Code-Review+2 Workflow+1

This last change was partly to fix an issue with the job trying to install with pip3 and pip3 not being present. That was done from within the script though. Maybe we need both to be safe.

We need to get this fixed though, and the current pip usage obviously isn't working right, so since this is fairly trivial and is blocking a lot of folks, I'm going to push this through. We can always revert if necessary.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1