Add neutron to stable upper-constraints
Some projects, notably neutron plugins, are using the neutron service as
a library. There is the neutron-lib package, but apparently they have
not moved everything needed over to there yet.

This is being added directly in stable branches only. Master needs to
get the latest version available, so these will need to be added at each
release so the stable branches use the correct version of neutron for
their respective stable release.

Change-Id: I8e95dc48b8ba2a7e16f1d34eb5ad0e87a3e96f4f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>