Switch to ussuri job templates
We don't really need to follow the project-wide cycle templates, but
most of the jobs defined in them are what we probably want to run
anyway. This drops the py2 job template, since all of our code is only
ever run under py3 now anyway, and switches from the train template to
the ussuri template.

Change-Id: Ic55fa1f88141e7f592317fe334643bd7d81b5fd9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>