Update patch set 1
Patch Set 1: Code-Review-1

There was a lot of useful conversation in the review. please add it to the commit message. 

The take away I have from reading the lifecycle here https://www.djangoproject.com/download/

is that we want to update the cap to < 3.0 (or < 3.2). depending on the decision on whether horizon wants to follow 2.2 LTS or not. 

So a new cap might be the better solution.

Patch-set: 1