Update patch set 1
Patch Set 1:

(1 comment)

this looks ok (with one note on the regex).  I think this should be expanded to allow for python2 being defined in global-reqs/upper-constraints but not in the project.  I don't think projects would be allowed to remove the python2 lines with just this patch as is.

Patch-set: 1
Label: Verified=0