Add horizon to upper-constraints
horizon plugins use horizon as a library.
Having horizon in upper-constraints would be helpful
to avoid unexpected breakage due to a newer horizon releases,
particularly in stable branches.

The reason we don't have horizon in upper-constraints so far is
because most horizon plugins would like to work and test with horizon
in a same branch, but Zuul v3 allows such tests.
This reasoning is no longer true.

Note that horizon has switched to cycle-with-intermediary release mode
in Ussuri so we can have horizon in u-c in the master branch.

Change-Id: I421ee2d498ea5100d9b52e356b46f2d594638b5b