Update patch set 2
Patch Set 2:

> the right way to do this is to have a horizon-lib library that the shared functions are stored within, I cannot approve this for the master branch (neutron has the same issue)

Theoretically I understand it, but I don't expect that happens in near future due to the development resources.
It is an issue of investment and perhaps it is a topic related to the board of director, the foundation and TC, so I have nothing to say.

Apart from the above, what should I do to fix gate failures in stable branches? Is it better to propose the same thing to stable/train directly? or to do nothing, or to cap horizon in stable branches?

Patch-set: 2