use enmerkar instead of django-babel
django-babel is unmaintained since 2017 and it is not compatible
with the latest Django releases like 2.2 or 3.
enmerkar [1][2] forked django-babel and it looks a good replacement.
Looking at the project history, the author looks active as commits
happened around recent Django releases.

Note that enmerkar requires Django>=2.2 which supports python3 only.

[1] https://github.com/Zegocover/enmerkar
[2] https://pypi.org/project/enmerkar/

Change-Id: I46f7ad26695f729eb2374265cf83e385ec7ce863