Update patch set 2
Patch Set 2: Code-Review+1

Many Thanks for this, Eyal

It may be that we need to blacklist 1.3.0.  I think the problem came up by an automated move of the version and not because someone qualified the new version of confluent-kafka-python, so some review of versions is in order.

Patch-set: 2
Reviewer: Gerrit User 10311 <10311@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1