Constraint openstackclient for stable branch
Openstackclient[1] does not have stable branches. This is problematic because
components that depends on it will break whenever it releases a newer version
with requirements which conflicting with its dependents in the stable branch.

This patch constraints the openstackclient version that is compatible with the
stable branch so we can bring back sanity and world peace.

1. https://github.com/openstack/openstackclient

Change-Id: I80c7f1029e7827ccb59a043a6c2bc557fb664d3f