Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

I think the issue is just adding openstackclient to upper-constraints, not necessarily that there aren't stable branches. The relevant part is that a given branch needs to have the upper constraint for the library version that was tested during that cycle so when later branches pick up a newer version (because it's an independently released deliverable) it will need to have a proposal raising the upper-constraint that is tested like we do for any of our other libraries.

So I think this is probably OK to propose directly to stable branches, but it really needs to be added to master and others as well.

Patch-set: 1