Update patch set 1
Patch Set 1: Code-Review-1

> ya, it seems we are getting quite a few packages that are
 > constrained on stable branches only, maybe this is a more general
 > issue?

I don't think that's the case here though. This isn't an issue of something that needs to be constrained on stable branches only. It's an issue that we are using a lib but it's not being tracked for some reason.

Though looking at this more, it's not tracked in global-requirements, so I don't see how someone is using this as a dependency. I don't believe it is supposed to be used that way. openstackclient is just a convenience wrapper on getting all python-openstackclient plugins installed. If a project needs some of those, it should be pulling in the plugins directly, not using this to pull in the world. So actually, I don't think we should do this. At least until we understand the whole story here.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1