Update patch set 1
Patch Set 1:

Further discussion started here:

http://eavesdrop.openstack.org/irclogs/%23openstack-requirements/%23openstack-requirements.2020-01-06.log.html#t2020-01-06T22:20:03

Summary is - the python-openstackclient repo is pulling in a job template defined in openstackclient. That repo and job are branchless, so they always refer to master. This results in any stable branch repos that are using this job template getting in requirements pulled in from master. This obviously breaks because that will end up getting newer versions of libs than what is in the stable branch's upper-constraints.

So since openstackclient is branchless, and the job itself not really being able to handle it, the "correct" approach (as far as we can tell so far) is to limit that job to only run on master.

Patch to do that is here - https://review.opendev.org/#/c/701293/

Patch-set: 1