demarcate python-memcached requirements
To support Python3, we need python-memcached 1.58 or above. However,
projects such as keystonemiddleware[1] have python-memcached 1.56
in its lower-constraints.txt. For stable/stein, Python3 is now the
defult[2]. However, we generally avoid changing lower constraints in
the stable branches so we don't inadvertently break the downstream
consumers. Therefore, we need to separate python-memcached requirement
for the respective python versions.

[1] https://github.com/openstack/keystonemiddleware/blob/stable/stein/lower-constraints.txt#L61
[2] https://governance.openstack.org/tc/goals/selected/stein/python3-first.html

NOTE: in addition to the original backport, this patch also cap
the gunicorn version to less than 20.0.0 as the later versions
no longer supports Python2. Without this change, the
requirements-integration gate will fail.

Change-Id: Ifb805954ff36420004d552a690046d8ddf3bed5c