Update patch set 3
Patch Set 3: Code-Review-1

See comments in bug report. This is not the correct way to fix stable/rocky.

We may still need this in master, but if we do, it at least needs to be the latest release of neutron-lib before the 2.0.0 release.

Patch-set: 3
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1