Update patch set 3
Patch Set 3:

> So maybe the issue is that neutron/tempest-plugin stable/rocky jobs
 > use the neutron/stable/rocky branch, but that has this in
 > requirements.txt:
 > 
 > requirements.txt:neutron-lib>=1.18.0 # Apache-2.0
 > 
 > So there's no cap there.
 > 
 > But trying to pin that triggers a requirements failure,
 > https://review.opendev.org/#/c/702979/
 > 
 > Which I think is the reason I started hacking in this repo.  What a

The job should be using the -c{CONSTRAINTS_URL} for stable/rocky, so that entry in requirements.txt should be fine. It should lock it down so 1.18.0 is the *only* version that will match the requirement.


 > mess py2 has been...

True, but this also is highlighting a latent bug in those jobs. It has not been testing what we've thought it's been testing for several releases now, so at least this is shining some light on that.

Patch-set: 3