Update patch set 1
Patch Set 1:

> > Not sure if we also need to cap this in global-requirements.txt,
 > > but this looks happy.
 > 
 > best follow-up but jobs I've seen so far consulted u-c

Yeah, this will make sure jobs using this upper-constraint will not get anything higher than what is set here, but my concern is the proposal bot may try to raise this upper-constraint if it not capped in the global-requirements file to explicitly tell it to keep these older versions limited to 1.1.0.

I can't recall - we'll see what Matthew knows about that.

Patch-set: 1