blacklist blockdiag 2.0.0
blockdiag has been used as an indirect dependcy via sphinx
for several releases. For the past few releases nova has monkey
patched blockdiag to around there global replacement of codec.getreader
under python 3.

The 2.0.0 release broke how that monkey patching worked and also
was the first release in several years. Now that the project is active
again we have reported the python 3 issue to the maintainer
https://github.com/blockdiag/blockdiag/issues/126 and it
has been adressed by https://github.com/blockdiag/blockdiag/pull/129
in 2.0.1 which has now been released.
https://pypi.org/project/blockdiag/2.0.1/

This change blacklists 2.0.0. This change does not raise the minium
requirement. Nova can selectively raise its minium requirement in the
future in tree if it chooses too remove support for monkey patching
older releases.

Other project are recommended to use 2.0.1 or later as this is the first
release which is python3 only and is compatiable with the sphinx
extentions we use without monkey patching.

Change-Id: I414d9e604616d49b4456ebc348418e9b1b09058b