blacklist blockdiag 2.0.0
blockdiag has been used as an indirect dependcy via sphinx
for several releases. For the past few releases nova has monkey
patched blockdiag to around there global replacement of codec.getreader
under python 3.

The  2.0.0 broke how that monkey patching worked and also
was the first release in several years. Now that the project is active
again we have reported the python 3 issue to the maintainer and it
has been adressed in 2.0.1 which has now been released.

This change blacklists 2.0.0 and cap it at <3.0 so that projects
will not be broken by the next major release. This change does
not raise the minium requirement however nova can now raise its
minium requirement to blockdiag>=2.0.1 if it chooes too to allow
removal of monkey patching should it desire.

Other project are recommended to use 2.0.1 or later as this is the first
release which is python3 only and is compatiable with the sphinx extentions
we use without monkey patching.

Change-Id: I414d9e604616d49b4456ebc348418e9b1b09058b