Update patch set 1
Patch Set 1:

> > yeah, I will not suggest to cap the requirement which can end up
 > > doing for many lib. its time to drop the py2 aggressivly. Same I
 > > mentioned in ML- http://lists.openstack.org/pipermail/openstack-discuss/2020-February/012356.html
 > 
 > Thanks @gmann @andreas for the comments, I am not sure how to move
 > forward but by looking at uc we have some many dependencies are
 > capped with py27 conditions

Yeah we did for few dependencies but now oslo and QA tooling droppping py2 so many/almost all projects still running py2 jobs will start failing. 

Next week is the deadline to complete the py2 drop work so projects failing on py2 jobs has to drop the py2 on priority. Tempest gate is also blocked due to that and we are on fixing it by dropping py2 jobs.

Patch-set: 1