Update patch set 2
Patch Set 2:

Ugh, apparently somewhere along the line I stopped testing that change with master and only did it with the branch that had my change to keystone. It is indeed broken right now.

I have an idea how we can get around this. I'll try to get a patch pushed today.

Patch-set: 2