Add pywinrm to global-requirements.txt
The "oswin-tempest-plugin" is using the pywinrm library, yet it's
not included in the global-requirements.txt file.

Although not set as as a hard dependency, the Manila Windows driver
(similar to the generic Linux driver but using Windows instances)
is also using pywinrm.

pywinrm allows sending commands to remote Windows hosts via WinRM,
pretty much resembling SSH.

Change-Id: I57f958ca93a2dd424ce988cd28816550bbd6e2d9
Closes-Bug: #1745823