Update patch set 1
Abandoned

Sorry for my late response, I abandon these for now since they are not necessary for now.

Else to anwser your questions, it's more for testing or scripting purpose than to be used in a client library.

I'll reopen (or recreate a similar) this patch when I'll be more ready with this topic.

Patch-set: 1
Status: abandoned