Update gitdb to 4.0.1, gitdb2 to 4.0.2
The gitdb 4.0.0 release and gitdb2 4.0.1 release has been deleted from
PyPi and (if I'm reading [1] correctly, which is a big *if*) has been
replaced with 4.0.1 and 4.0.2 respectively.  [1]:

  gitdb v4.0.1 and gitdb2 v4.0.2 should be functionally equivalent to
  gitdb2 v3.0.3 with gitdb v4.0.2 simply installing gitdb >= 4.0.1 and
  with gitdb v4.0.1 requiring smmap rather than smmap2
  (gitpython-developers/smmap#44).

Additionally, it seems that gitdb now also requires smmap>=3.0.1.
This project is in a similar renaming phase, but 3.0.1 seems to be the
correct place to be [2]:

  v3.0.1 for both packages should be functionally equivalent to smmap2
  v2.0.5, with smmap2 v3.0.1 simply installing smmap >= 3.0.1.

[1] https://github.com/gitpython-developers/gitdb/issues/59
[2] https://github.com/gitpython-developers/smmap/issues/44

Change-Id: I268fb7e1dc753c82983736572b3df2f152f8cf51