Update patch set 2
Patch Set 2: Code-Review-1

Skipping just this version won't resolve the problem -- it's not a bug in setuptools, they've literally removed code I believe? We need to fix the other libs that are having problems, because the second setuptools==46.0.1 is released, this problem reappears, which could even be minutes/hours? I don't see the point of merging this as-is.

Patch-set: 2
Reviewer: Gerrit User 10273 <10273@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1