Update patch set 3
Patch Set 3:

yes we have discussed pytest in the context of test runners in the past.

currently horizon is using django test so i would guess its already not using subunit protocal to recored the results so swapping to pytest would not cnage that in horizons case.

but also they want to use the other part of pytest, the test framework to work around the non deterministic way the test are executed in so thttps://governance.openstack.org/tc/reference/pti/python.html#python-test-running is only related to part of this change it is not relvent for the test framework part.

anyway i dont really mind too much one way or another im jsut trying to play devils advocate here but the way stestr uses the subunit protocol has been cause issue for nova for 2-3 release now when there is too much log output.


anyway i don't think this is unreasonable to add even if we don't encourage other project to adopt it.

im not sure if it still work by the way but 
https://pypi.org/project/pytest-subunit/
was ment to output the test result in subunit format although its not maintained anymore.

Patch-set: 3
Reviewer: Gerrit User 11604 <11604@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0