Update patch set 1
Patch Set 1:

It's still not clear to me why this cannot wait for the bot to update the constraint.

You state that 1.3.6 does not require a bindep, should it?  If yes that may be a reason to update to 1.3.7 but I still need to know why it's so urgent (is a project blocked?)

Patch-set: 1
Label: Verified=0