[ocata only] Add stestr to upper constraints
Stestr dropped python 2 support with version 3.0.0, but setup.cfg was
wrongly configured in that release [1], which was fixed in version
3.0.1. Unfortunately this causes issues in python 2 as pip falsely
installs 3.0.0 as it is not restricted as python 3 only yet.

In Ocata there isn't any constraint for stestr, but grenade job on
Pike installs Tempest with using upper-constraints.txt from Ocata,
which does not cap stestr and so pip installs version 3.0.0, and
grenade job fails with the error when running stestr:

ImportError: No module named configparser

Either stestr 3.0.0 needs to be backlisted or needs to be constrained.
This patch sets the stestr===1.1.0 upper constraint.

[1] https://github.com/mtreinish/stestr/issues/292

Change-Id: I7f372e977d3c61f1a680866e10b2b7a429a99ae3