Update patch set 1
Patch Set 1:

It's not in anyone's requirements files, not setup.cfg. looks like the only place it's installed is through a devstack script.

I think for now, short term answer is for us to know not to remove it. Long term, I think the projects that are using it need to explicitly as it to their requirements, our as an extra in setup.cfg.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0