Update patch set 1
Patch Set 1:

So I think we may have a problem here. We need to raise dnspython to allow the current consumers of dnspython3 to move to it. But the last version of dnspython3 has a requirement for the older version of dnspython.

We might need to temporarily break designate (the only consumer of this) by skipping ahead to the step of removing dnspython3 and raising dnspython. Then that could be quickly followed by patches for the two designate repos that have this to switch over to dnspython.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0