Update patch set 7
Patch Set 7:

So looks like we are royally screwed at the moment. There's no way we can get past the 'legacy-tempest-dsvm-neutron-full' gate without stestr minimal version 1.0.0 as that's the version where the stestr module starting to export the 'run_command' which tempest dependes on. However, with stestr version 1.0.0, it requires pbr version 2.0.0 or greater. This will fail the 'requirements-tox-py27-check-uc' gate as the upper constraint for pbr is 1.0.0. Besides, the 'requirements-integration' gate is also failing with weird directory permission issue.

Patch-set: 7