Update patch set 2
Patch Set 2:

> > I'm kinda thinking that ocata might be more in 'archival mode'
 > 
 > Agreed. That means we'll also need to disable any grenade-related
 > jobs (i.e. neutron-grenade) in stable/pike if we want to land
 > anything in stable/pile. That's unless stable/pike is also in
 > archival mode.

The commitment for extended maintenance branches was that we would continue to test as best we can. So I think it makes sense to get rid of some of the jobs like grenade if it becomes too difficult to keep them working right. As long as the unit tests continue to pass.

Not sure if that would be an issue for pike. But pretty much rocky and later are in extended maintenance, so I guess we can see where we can draw the line.

Patch-set: 2