Update patch set 1
Patch Set 1:

> Was a bug filed upstream to confirm the behavior?  What is the plan
 > with dealing with the new paths going forward (aka, when can this
 > patch be reverted)?

As of now we are not sure if it is setuptools bug and they can fix it or we need to update devstack for the 'pip install -e'.

For now, I have filled devstack (critical) bug for further debugging.  https://bugs.launchpad.net/devstack/+bug/1886237

As part of that bug (either we fix in devstack or setuptools fix it), we can revert this.

Patch-set: 1
Reviewer: Gerrit User 8556 <8556@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0