Update patch set 4
Patch Set 4:

this isn't in global-requirements, and wasn't in upper-constraints...  this is a dep of a linter (and ONLY a dep of a linter).

this makes me think we should add it to blacklist.txt and have downstreams set their own required versions, like with pylint itself.

all comments are removed when upper-constraints are regenrated, but at least it is a bigger block of text to see in the resulting diff

Patch-set: 4
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0