Update patch set 2
Patch Set 2:

by the way im not +1 or -1 on this change becaseu i know my personal view differes from the requiremtn tesms view.

i kind of feel like we should be caping the major version by default and updating them periodically.

i also know the rational for not doing that and following head. that said when an indriect dependcy announces a major version change like this and breaks one of our direct dependincies and we dont assertivly prevent it happening again it feels like we are repeating the same mistake.

the major downside with capping is we need to activly reasses it and make sure we dont fall too far behind so that has problems too.

Patch-set: 2