Update patch set 2
Patch Set 2:

It would be interesting to understand how the cross job for Nova did not catch this.

We "reassess" our limits with every update to upper-constraints. So we shouldn't really need to cap major versions since we implicitly cap *every* version. The testing usually is really good at catching these kinds of issues, but apparently in this case, none of the test cases run must hit this code path.

Patch-set: 2