Update patch set 2
Patch Set 2:

> (2 comments)

Hi Matthew, oslo.cache is the client server to acess the memcached server, if the memcached has sasl enabled，oslo.cache cannot have the authority to access the memcached, because we need the user name and password. Therefore we can access the memcached via the python-binary-memcached package. we only have python-memcached package in current oslo.cache, that's not enough. If we want to support sasl even tls protocol，we may add this package to default.

Patch-set: 2