Update patch set 4
Patch Set 4:

why not think about using the python-binary-memcached instead of python-memcached, the former has richer function than python-memcached. However, having the two package in the project is also another a good choice. We also can make a better choice among the python-memcached, python-binary-memcached and pymemcache with the improvement of functions in the future.

Patch-set: 4