Update patch set 4
Patch Set 4:

Looks like pymemcache does not support the binary protocol, so does not support sasl.  (there's another issue for missing ipv6 support)

https://github.com/pinterest/pymemcache/issues/54

looks like this is required for sasl support, sadly  In global-requirements, do you mind adding a comment above the entry stating that it is only required for sasl/binary protocol (so that it can be cleaned up if/when sasl/binary protocol is merged into another 'approved' library).

Patch-set: 4