Update patch set 1
Patch Set 1: Code-Review+1

yep looks good ill comment when i have filed the os-vif bug and fixed it. this will fail until then so you will need to add a depends on to the os-vif patch.

its a trivial fix so no need to revert the upper-constraint bump for pyroute2.

Patch-set: 1