Update rbd-iscsi-client to 0.1.8
This patch updates the upper-constraints.txt to
accomodate the new version of rbd-iscsi-lient for
the cinder ceph-iscsi driver.

This has a new required parameter to delete_disk
that allows the preservation of the backing rbd image,
to ensure cleanup of the iscsi gateway on nova detach
without deleting the rbd image from the pool.

Change-Id: I48c11e441afca3dea1fa69500e9b51764b3730d3