Update patch set 1
Patch Set 1:

note starting with this change we have started to see som e tempest failures with 
"Multiple possible networks found, use a Network ID to be more specific" 

http://logstash.openstack.org/#/dashboard/file/logstash.json?query=message:%5C%22Multiple%20possible%20networks%20found,%20use%20a%20Network%5C%22

set the time span to 30days.

its not clear that this is related to this change but just an fyi that it might be.

it could also just be a case that soemthing merged in tempest or devstack to cause this around the same time an this just happened to hit it

Patch-set: 1