Add CONFspirator to requirements
CONFspirator is a config library inspired by oslo.config
built to solve somewhat different problems.

It is license compatible, py3 only, written by the Adjutant
PTL, and will be maintained along with Adjutant, as well
as hopefully as a separate library going forward.

It is already in Adjutant as of:
https://opendev.org/openstack/adjutant/commit/c750fd6d6c0ab5390d66428069a4752dfaae1182

Docs:
https://confspirator.readthedocs.io/
Code:
https://gitlab.com/catalyst-cloud/confspirator

Change-Id: Ibdb8785b3fa6234c8fadbeaa3a672eda373c7707