Update patch set 1
Patch Set 1:

Should we revert this now for consistency reason, as we don't need it anymore? Wheel 0.35.1 was released [1] which (maybe as a side effect) solves the issue for us [2] (changelog is not up-to-date, but at least it shows the relevant change).

Let me know about what do you think and if you say so then I'll push a revert patch for this!

[1] https://pypi.org/project/wheel/#history
[2] https://wheel.readthedocs.io/en/stable/news.html

Patch-set: 1