Update patch set 2
Patch Set 2:

Hi, yes I expected as much with the error.
The use case is the following: Devstack on centos7 requires a recent version of setuptools, however it is not upgraded as the base image contains a very old version. 
What do you think is the best course of action?

Patch-set: 2