bump pecan to 1.4.0
Neutron recently lifted pecan to 1.4.0 to fix [0], to keep neutron
stadium projects aligned pecan should be bumbed in global requirements
as well.
This patch is more a kind of survey to see how we van progress with it.
The issue which I am aware with pecan 1.4.0 is [1], and resulted in
baning pacan in [3].

[0] https://bugs.launchpad.net/neutron/+bug/1829042
[1] https://storyboard.openstack.org/#!/task/40652
[3] https://review.opendev.org/746261

Closes-Bug: #1894864
Change-Id: Ied3e806102e0100bdbab7b29209d8eff5b003bdd