Update patch set 1
Patch Set 1:

The rationale behind these changes LGTM.

However I've some concerns related to pbr:
- pbr rely on setuptools [1] and still support python2.7 [2]
- setuptools 50.3.0 only support python3 [3]

So I wonder if we should also define a version which support python2.7 to avoid issues on with this context. setuptools dropped the support of python 2 with 45.0.0 [4] so we could use the version 44.1.1 [5] for this use case.

[1] https://opendev.org/openstack/pbr/src/branch/master/setup.py#L16
[2] https://opendev.org/openstack/pbr/src/branch/master/setup.cfg#L25
[3] https://pypi.org/project/setuptools/50.3.0/
[4] https://setuptools.readthedocs.io/en/latest/history.html#v45-0-0
[5] https://pypi.org/project/setuptools/44.1.1/

Patch-set: 1