Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> do you have a link to a bug tracking whitebox-tempest-plugin's need for this?

It's already in the code - we started using it without permission, so to speak. The reason this patch came up now is we want to have whitebox packaged as an RPM to run TripleO jobs in our gate, but to get our RPM package we need our dependencies packaged as well. And to get our dependencies packaged, we need them listed in global-requirements.txt.

> 
> other than that, ok

Patch-set: 3