Update patch set 4
Patch Set 4:

Let's drop the crudini requirement - whitebox only calls it on the command line anyways, not as a Python import. For our RPM needs, we'll either add it as a dependencies without a minimum version to use the older version that CentOS has, or just not mention it at all and document the need to install it manually.

Patch-set: 4