DNM: test horizon integration tests with the previous jquery-migrate
horizon integration test started to fail since Jan 28.
xstatic-jquery-migrate in u-c was updated at Jan 28 and
it was just before the test failure started.
I wonder xstatic-jquery-migrate upgrade triggers the horizon
integration test failure.
This is to verify what happens in the horizon integration tests
with the previous version of xstatic-jquery-migrate.

Change-Id: I617e3b73d93792ce23a3c2427db45764dac85498