Update patch set 2
Patch Set 2:

Yeah, removing caps is a tricky and annoying process. I actually can't remember offhand how that works. We really should capture that in a doc somewhere.

I seem to recall it being a multi-step process. We may be able to find past instances of this if we look through the commit history. Will need to look, but I think you will actually need to break out removing the cap as a first step, then raising the upper constraints after that.

Patch-set: 2
CC: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>