Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could it be because https://review.opendev.org/c/openstack/oslo.policy/+/773947 is with python 3.8?

Yes, that's the issue. Neutron has a custom policy check type, FieldCheck, that stores a 'regex' attribute [1]. It's not possible to deepcopy regex and match objects before Python 3.7 [2][3]. I'm guessing the gate runs on Python 3.8.

Question is how to resolve this now. Would a fix in neutron be acceptable? I think it's reasonable, personally.

[1] https://github.com/openstack/neutron/blob/ab8a17c6a49/neutron/policy.py#L377
[2] https://docs.python.org/3/whatsnew/3.7.html#re
[3] https://stackoverflow.com/a/52885510/613428

Patch-set: 1
CC: Gerrit User 15334 <15334@4a232e18-c5a9-48ee-94c0-e04e7cca6543>