Add pycdlib to global-requirements.txt
pycdlib is a pure python library used to parse,
read and write ISO9660 files. With LGPLv2 as the
license it confirms to openstack licensing guidelines.

This is an actively maintained library. As of this
writing the last change to this library was 8 days
ago and the first commit was over 6 years ago.

This library is compatible with python3
https://clalancette.github.io/pycdlib/python-compatibility.html
It is a well documented library and the documentation
is available at https://clalancette.github.io/pycdlib

This library will be used by ironic to read config
drive and write them to disk on the filesystem.

Change-Id: Idfb40ba12893d75731542044d79532ee27c2a499