Update patch set 1
Patch Set 1:

Where we consistently hit this issue is the dsvm-functional-py27 job in stable/train. Essentially the bug in OVS is that notifications are sent *as a transaction is being processed for each row* instead of for each row *after the transaction is committed*. This is already bad in that if a multi-row transaction failed in one of the later rows, we'd still get notified of rows making it that would actually be rolled back. Just for that reason, I'd say this was important.

But where we frequently hit this in the py27 job is that we send multi-row transactions with foreign key relationships with both rows being created at the same time. Due to the fact that tables are stored in a dict, and py27 doesn't maintain insert order like py3, it ends up hashing in a way that is *always* wrong based on the name of the table, i.e. the table referencing the row via foreign key get a notification sent that happens *before* the row is created that it references, so that field is sent as blank, breaking the parsing of the event where decisions are made based on that field.

There are other race conditions related to notifications due to the Row object being mutable and passed to an event processor in another thread. We can fix this by making that row an immutable copy of the object, but that means that we *never* will pass for the case where those foreign key rows are created simultaneously in py27, whereas before we'd get lucky a lot of the time because the transaction would complete before the notification thread managed to access the missing row.

Now, with all of the good reasons out of the way, I'd like to mention that I think the policy to generally not bump should be changed because semver is great and bugfixes are good. :D

Patch-set: 1